update workflows and relic points on profile routes
