+++ .github/workflows/deploy_production.yml
@@ -106,7 +106,8 @@ jobs:
           --build-arg GTM_TRACKING_ID=${{ secrets.GTM_TRACKING_ID }} \
           --build-arg FF_FULL_LOCKDOWN_ENABLED=false \
           --build-arg FF_GENERIC_BANNER_ENABLED=false \
-          --build-arg FF_INCIDENT_BANNER_ENABLED=false
+          --build-arg FF_INCIDENT_BANNER_ENABLED=false \
+          --build-arg RELIC_API_URL=${{ secrets.RELIC_API_URL }}
           docker tag portal ${{ env.ECR_IMAGE }}
           docker push ${{ env.ECR_IMAGE }}
           rm privy_verification_key.pem
