+++ apps/portal/Dockerfile
@@ -28,6 +28,7 @@ ARG FF_FULL_LOCKDOWN_ENABLED=${FF_FULL_LOCKDOWN_ENABLED}
 ARG FF_INCIDENT_BANNER_ENABLED=${FF_INCIDENT_BANNER_ENABLED}
 ARG FF_GENERIC_BANNER_ENABLED=${FF_GENERIC_BANNER_ENABLED}
 ARG VITE_DEPLOY_ENV=${VITE_DEPLOY_ENV}
+ARG RELIC_API_URL=${RELIC_API_URL}
 
 ENV ALCHEMY_MAINNET_API_KEY=${ALCHEMY_MAINNET_API_KEY}
 ENV ALCHEMY_API_KEY=${ALCHEMY_API_KEY}
@@ -57,7 +58,7 @@ ENV FF_FULL_LOCKDOWN_ENABLED=${FF_FULL_LOCKDOWN_ENABLED}
 ENV FF_INCIDENT_BANNER_ENABLED=${FF_INCIDENT_BANNER_ENABLED}
 ENV FF_GENERIC_BANNER_ENABLED=${FF_GENERIC_BANNER_ENABLED}
 ENV VITE_DEPLOY_ENV=${VITE_DEPLOY_ENV}
-
+ENV RELIC_API_URL=${RELIC_API_URL}
 WORKDIR /app
 
 COPY package.json \
