+++ apps/portal/app/routes/app+/quest+/index.tsx
@@ -17,133 +17,26 @@ import {
 import { ErrorPage } from '@components/error-page'
 import ExploreHeader from '@components/explore/ExploreHeader'
 import { PointsEarnedCard } from '@components/points-card/points-card'
-import { QuestSetCard } from '@components/quest/quest-set-card'
 import { QuestSetProgressCard } from '@components/quest/quest-set-progress-card'
 import { ReferralCard } from '@components/referral-card/referral-card'
 import RelicPointCard from '@components/relic-point-card/relic-point-card'
+import { useRelicCounts } from '@lib/hooks/useRelicCounts'
 import { calculatePointsFromFees, invariant } from '@lib/utils/misc'
 import { LoaderFunctionArgs } from '@remix-run/node'
-import { Await, Link, useLoaderData } from '@remix-run/react'
+import { Await, useLoaderData } from '@remix-run/react'
 import { fetchWrapper } from '@server/api'
 import { requireUserWallet } from '@server/auth'
 import { getQuestsProgress } from '@server/quest'
 import {
   BLOCK_EXPLORER_URL,
-  COMING_SOON_QUEST_SET,
   HEADER_BANNER_HELP_CENTER,
-  QUEST_LOG_DESCRIPTION,
   STANDARD_QUEST_SET,
 } from 'app/consts'
 import { isAddress } from 'viem'
 
-interface GraphQLResponse<T> {
-  data?: T
-  errors?: Array<{ message: string }>
-}
-
-const GetMintCountDocument = {
-  query: `
-    query GetMintCountUntilDate($address: String!, $cutoff_timestamp: Int!) {
-      voucherRedeemedEvents(
-        where: { redeemer: $address, timestamp_lte: $cutoff_timestamp }
-      ) {
-        totalCount
-      }
-    }
-  `,
-} as const
-
-interface GetMintCountQuery {
-  voucherRedeemedEvents: {
-    totalCount: number
-  }
-}
-
-interface GetMintCountQueryVariables {
-  address: string
-  cutoff_timestamp: number
-}
-
-async function fetchGraphQL<T, V>(
-  document: { query: string },
-  variables: V,
-): Promise<GraphQLResponse<T>> {
-  const endpoint = process.env.RELIC_GRAPHQL_ENDPOINT
-  if (!endpoint) {
-    throw new Error('RELIC_GRAPHQL_ENDPOINT not configured')
-  }
-
-  const response = await fetch(endpoint, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      query: document.query,
-      variables,
-    }),
-  })
-
-  if (!response.ok) {
-    throw new Error(`GraphQL request failed: ${response.statusText}`)
-  }
-
-  return response.json()
-}
-
-const GetRelicHoldingsDocument = {
-  query: `
-    query GetRelicHoldings($address: String!) {
-      account(address: $address) {
-        tokens {
-          totalCount
-        }
-        voucherRedeemedEvents {
-          totalCount
-        }
-      }
-    }
-  `,
-} as const
-
-interface GetRelicHoldingsQuery {
-  account: {
-    tokens: {
-      totalCount: number
-    }
-    voucherRedeemedEvents: {
-      totalCount: number
-    }
-  }
-}
-
-interface GetRelicHoldingsQueryVariables {
-  address: string
-}
-
 export async function loader({ request }: LoaderFunctionArgs) {
   const userWallet = await requireUserWallet(request)
   invariant(userWallet, 'Unauthorized')
-
-  const cutoffTimestamp = 1735516799
-
-  const mintCountData = await fetchGraphQL<
-    GetMintCountQuery,
-    GetMintCountQueryVariables
-  >(GetMintCountDocument, {
-    address: userWallet,
-    cutoff_timestamp: cutoffTimestamp,
-  })
-
-  const holdingsData = await fetchGraphQL<
-    GetRelicHoldingsQuery,
-    GetRelicHoldingsQueryVariables
-  >(GetRelicHoldingsDocument, {
-    address: userWallet,
-  })
-
-  const mintCount = mintCountData.data?.voucherRedeemedEvents?.totalCount ?? 0
-  const holdCount = holdingsData.data?.account?.tokens?.totalCount ?? 0
 
   const userProfile = await fetchWrapper(request, {
     method: UsersService.getUserByWalletPublic,
@@ -179,18 +72,13 @@ export async function loader({ request }: LoaderFunctionArgs) {
     userProfile,
     userTotals,
     inviteCodes: inviteCodes.invite_codes,
-    relicHoldCount: holdCount.toString(),
-    mintCount,
   }
 }
 
 export default function Quests() {
-  const { userTotals, inviteCodes, relicHoldCount, details, mintCount } =
+  const { userTotals, inviteCodes, userWallet, details } =
     useLoaderData<typeof loader>()
-
-  const nftMintPoints = mintCount * 2000000
-  const nftHoldPoints = parseInt(relicHoldCount) * 250000
-  const totalNftPoints = nftMintPoints + nftHoldPoints
+  const { mintCount, holdCount, totalNftPoints } = useRelicCounts(userWallet)
 
   return (
     <div className="p-10 w-full max-w-7xl mx-auto flex flex-col gap-5 max-md:p-5 max-sm:p-2">
@@ -272,7 +160,7 @@ export default function Quests() {
             </div>
             <RelicPointCard
               relicsMintCount={mintCount}
-              relicsHoldCount={parseInt(relicHoldCount)}
+              relicsHoldCount={holdCount}
               relicsPoints={totalNftPoints}
             />
           </div>
@@ -293,45 +181,6 @@ export default function Quests() {
               )}
             </Await>
           </Suspense>
-        </div>
-        <div className="flex flex-col gap-10 max-md:gap-5">
-          <div className="space-y-5 max-md:space-y-3">
-            <Text variant="headline">Quest Log</Text>
-            <Text variant="body" className="text-foreground/70">
-              {QUEST_LOG_DESCRIPTION}
-            </Text>
-          </div>
-          <ul className="grid grid-cols-1 gap-10 sm:grid-cols-1 md:grid-cols-1 lg:grid-cols-2 max-md:gap-5">
-            <Suspense fallback={<Skeleton className="h-full w-full" />}>
-              <Await resolve={details}>
-                {(resolvedDetails) => (
-                  <Link to={STANDARD_QUEST_SET.navigatePath} prefetch="intent">
-                    <li className="col-span-1 h-full">
-                      <QuestSetCard
-                        imgSrc={STANDARD_QUEST_SET.imgSrc}
-                        title={STANDARD_QUEST_SET.title}
-                        description={STANDARD_QUEST_SET.description}
-                        numberQuests={resolvedDetails.numQuests}
-                        numberCompletedQuests={
-                          resolvedDetails.numCompletedQuests
-                        }
-                      />
-                    </li>
-                  </Link>
-                )}
-              </Await>
-            </Suspense>
-            <li className="col-span-1 h-full">
-              <QuestSetCard
-                disabled
-                imgSrc={COMING_SOON_QUEST_SET.imgSrc}
-                title={COMING_SOON_QUEST_SET.title}
-                description={COMING_SOON_QUEST_SET.description}
-                numberQuests={0}
-                numberCompletedQuests={0}
-              />
-            </li>
-          </ul>
         </div>
       </div>
     </div>
